cleanup. component file
