+++ client/signup/signup-header/style.scss
@@ -51,7 +51,7 @@
 	}
 }
 
-.signup-header__banner-refund {
+.signup-header__overflow-banner {
 	top: 0;
 	left: 0;
 	margin-top: 0;
